Refactor Swap to use common ProofGenerationForm component
