--- client/src/components/Swap/OnRamperProof.tsx
@@ -1,378 +1,0 @@
-import React, { useContext, useState } from 'react';
-import { useAsync, useUpdateEffect } from "react-use";
-import { ArrowLeft } from 'react-feather';
-import styled from 'styled-components';
-
-import { Button } from "../Button";
-import { TitleCenteredRow } from '../layouts/Row'
-import { Col } from "../legacy/Layout";
-import { ThemedText } from '../../theme/text'
-import { LabeledTextArea } from '../legacy/LabeledTextArea';
-import { ProgressBar } from "../legacy/ProgressBar";
-import { NumberedStep } from "../common/NumberedStep";
-import { LabeledSwitch } from "../common/LabeledSwitch";
-import { DragAndDropTextBox } from "../common/DragAndDropTextBox";
-import ProofGenSettingsContext from '../../contexts/ProofGenSettings/ProofGenSettingsContext';
-
-import { downloadProofFiles, generateProof } from "../../helpers/zkp";
-import { insert13Before10 } from "../../scripts/generate_input";
-import { PLACEHOLDER_EMAIL_BODY } from "../../helpers/constants";
-// import { packedNBytesToString } from "../helpers/binaryFormat";
-import {
-  INPUT_MODE_TOOLTIP,
-  PROVING_TYPE_TOOLTIP
-} from "../../helpers/tooltips";
-
-const generate_input = require("../../scripts/generate_input");
-
-
-interface OnRamperProofProps {
-  loggedInWalletAddress: string;
-  setSubmitOrderProof: (proof: string) => void;
-  setSubmitOrderPublicSignals: (publicSignals: string) => void;
-  handleBackClick: () => void;
-}
- 
-export const OnRamperProof: React.FC<OnRamperProofProps> = ({
-  loggedInWalletAddress,
-  setSubmitOrderProof,
-  setSubmitOrderPublicSignals,
-  handleBackClick
-}) => {
-  /*
-   * Contexts
-   */
-  const {
-    isProvingTypeFast,
-    isInputModeDrag,
-    setIsProvingTypeFast,
-    setIsInputModeDrag
-  } = useContext(ProofGenSettingsContext);
-  
-  /*
-   * State
-   */
-  const [emailFull, setEmailFull] = useState<string>(localStorage.emailFull || "");
-
-  const [displayMessage, setDisplayMessage] = useState<string>("Generate Proof");
-  const [downloadProgress, setDownloadProgress] = useState<number>(0);
-
-  const [status, setStatus] = useState<
-    | "not-started"
-    | "generating-input"
-    | "downloading-proof-files"
-    | "generating-proof"
-    | "error-bad-input"
-    | "error-failed-to-download"
-    | "error-failed-to-prove"
-    | "done"
-    | "sending-on-chain"
-    | "sent"
-  >("not-started");
-
-  const [stopwatch, setStopwatch] = useState<Record<string, number>>({
-    startedDownloading: 0,
-    finishedDownloading: 0,
-    startedProving: 0,
-    finishedProving: 0,
-  });
-
-  const recordTimeForActivity = (activity: string) => {
-    setStopwatch((prev) => ({
-      ...prev,
-      [activity]: Date.now(),
-    }));
-  };
-
-  const filename = "circuit";
-
-  var Buffer = require("buffer/").Buffer; // note: the trailing slash is important!
-
-  // computed state
-  const { value, error } = useAsync(async () => {
-    try {
-      const circuitInputs = await generate_input.generate_inputs(
-        Buffer.from(atob(emailFull)),
-        "1",                                                        // TODO: Update me
-        "1"                                                         // TODO: Update me
-      );
-      return circuitInputs;
-    } catch (e) {
-      return {};
-    }
-  }, [emailFull, loggedInWalletAddress]);
-
-  // local storage stuff
-  useUpdateEffect(() => {
-    if (value) {
-      if (localStorage.emailFull !== emailFull) {
-        console.info("Wrote email to localStorage");
-        localStorage.emailFull = emailFull;
-      }
-    }
-  }, [value]);
-
-  if (error) console.error(error);
-  
-  const circuitInputs = value || {};
-  // console.log("Circuit inputs:", circuitInputs);
-
-  /*
-   * Handlers
-   */
-  const handleProvingTypeChanged = (checked: boolean) => {
-    if (setIsProvingTypeFast) {
-      setIsProvingTypeFast(checked);
-    }
-  };
-
-  const handleEmailInputTypeChanged = (checked: boolean) => {
-    if (setIsInputModeDrag) {
-      setIsInputModeDrag(checked);
-    }
-  };
-
-  return (
-    <Container>
-      <TitleCenteredRow style={{ paddingBottom: '1.5rem' }}>
-        <button
-          onClick={handleBackClick}
-          style={{ background: 'none', border: 'none', cursor: 'pointer' }}
-        >
-          <StyledArrowLeft/>
-        </button>
-
-        <ThemedText.HeadlineSmall style={{ flex: '1', margin: 'auto', textAlign: 'center' }}>
-          Complete Swap
-        </ThemedText.HeadlineSmall>
-
-        <LabeledSwitch
-          switchChecked={isProvingTypeFast ?? true}
-          onSwitchChange={handleProvingTypeChanged}
-          checkedLabel={"Speed"}
-          uncheckedLabel={"Privacy"}
-          helperText={PROVING_TYPE_TOOLTIP}
-        />
-      </TitleCenteredRow>
-
-      <Body>
-        <NumberedStep>
-          Open your payment sent Venmo transaction email and select 'Show original' to view the full contents. Download and drag
-          the .eml file into the box below or paste the contents directly.
-        </NumberedStep>
-        <OnRamperProofFormBodyTitleContainer>
-          <HeaderContainer>
-            <Title>Email</Title>
-            <LabeledSwitch
-              switchChecked={isInputModeDrag ?? true}
-              onSwitchChange={handleEmailInputTypeChanged}
-              checkedLabel={"Drag"}
-              uncheckedLabel={"Paste"}
-              helperText={INPUT_MODE_TOOLTIP}
-            />
-          </HeaderContainer>
-          {isInputModeDrag ? (
-            <DragAndDropTextBox
-              onFileDrop={(file: File) => {
-                const reader = new FileReader();
-                reader.onload = (e) => {
-                  if (e.target) {
-                    setEmailFull(e.target.result as string);
-                  }
-                };
-                reader.readAsText(file);
-              }}
-            />
-          ) : (
-            <LabeledTextArea
-              label=""
-              value={emailFull}
-              placeholder={PLACEHOLDER_EMAIL_BODY}
-              onChange={(e) => {
-                setEmailFull(e.currentTarget.value);
-              }}
-            />
-          )}
-        </OnRamperProofFormBodyTitleContainer>
-        <ButtonContainer>
-          <Button
-            disabled={emailFull.length === 0}
-            onClick={async () => {
-              console.log("Generating proof...");
-              setDisplayMessage("Generating proof...");
-              setStatus("generating-input");
-
-              const formattedArray = await insert13Before10(Uint8Array.from(Buffer.from(emailFull)));
-
-              // Due to a quirk in carriage return parsing in JS, we need to manually edit carriage returns to match DKIM parsing
-              console.log("formattedArray", formattedArray);
-              console.log("buffFormArray", Buffer.from(formattedArray.buffer));
-              console.log("buffFormArray", formattedArray.toString());
-
-              let input = "";
-              try {
-                input = await generate_input.generate_inputs(
-                  Buffer.from(formattedArray.buffer),
-                  "1",                                                        // TODO: Update me
-                  "1"                                                         // TODO: Update me
-                );
-              } catch (e) {
-                console.log("Error generating input", e);
-                setDisplayMessage("Prove");
-                setStatus("error-bad-input");
-                return;
-              }
-              console.log("Generated input:", JSON.stringify(input));
-
-              // Insert input structuring code here
-              // const input = buildInput(pubkey, msghash, sig);
-              // console.log(JSON.stringify(input, (k, v) => (typeof v == "bigint" ? v.toString() : v), 2));
-
-              /*
-                Download proving files
-              */
-              console.time("zk-dl");
-              recordTimeForActivity("startedDownloading");
-              setDisplayMessage("Downloading compressed proving files... (this may take a few minutes)");
-              setStatus("downloading-proof-files");
-              await downloadProofFiles(filename, () => {
-                setDownloadProgress((p) => p + 1);
-              });
-              console.timeEnd("zk-dl");
-              recordTimeForActivity("finishedDownloading");
-
-              /*
-                Generate proof
-              */
-              console.time("zk-gen");
-              recordTimeForActivity("startedProving");
-              setDisplayMessage("Starting proof generation... (this will take 6-10 minutes and ~5GB RAM)");
-              setStatus("generating-proof");
-              console.log("Starting proof generation");
-              // alert("Generating proof, will fail due to input");
-
-              const { proof, publicSignals } = await generateProof(input, "circuit"); 
-              console.log("Finished proof generation");
-              console.timeEnd("zk-gen");
-              recordTimeForActivity("finishedProving");
-
-              /*
-                Set proof
-              */
-              setSubmitOrderProof(JSON.stringify(proof));
-
-              /*
-                Retrieve public signals
-              */
-              // let kek = publicSignals.map((x: string) => BigInt(x));
-              // let soln = packedNBytesToString(kek.slice(0, 12));
-              // let soln2 = packedNBytesToString(kek.slice(12, 147));
-              // let soln3 = packedNBytesToString(kek.slice(147, 150));
-              // setPublicSignals(`From: ${soln}\nTo: ${soln2}\nUsername: ${soln3}`);
-              
-              /*
-                Set public signals
-              */
-              setSubmitOrderPublicSignals(JSON.stringify(publicSignals));
-
-              if (!circuitInputs) {
-                setStatus("error-failed-to-prove");
-                return;
-              }
-              setDisplayMessage("Finished computing ZK proof");
-              setStatus("done");
-              try {
-                (window as any).cJson = JSON.stringify(circuitInputs);
-                console.log("wrote circuit input to window.cJson. Run copy(cJson)");
-              } catch (e) {
-                console.error(e);
-              }
-            }}
-          >
-            {displayMessage}
-          </Button>
-        </ButtonContainer>
-        {displayMessage === "Downloading compressed proving files... (this may take a few minutes)" && (
-            <ProgressBar width={downloadProgress * 10} label={`${downloadProgress} / 10 items`} />
-          )}
-          <ProcessStatus status={status}>
-            {status !== "not-started" ? (
-              <div>
-                Status:
-                <span data-testid={"status-" + status}>{status}</span>
-              </div>
-            ) : (
-              <div data-testid={"status-" + status}></div>
-            )}
-            <TimerDisplay timers={stopwatch} />
-          </ProcessStatus>
-      </Body>
-    </Container>
-  );
-};
-
-const Container = styled.div`
-  width: 100%;
-  gap: 1rem;
-`;
-
-const Body = styled(Col)`
-  gap: 0.75rem;
-`;
-
-const OnRamperProofFormBodyTitleContainer = styled(Col)`
-  gap: 0rem;
-`;
-
-const ProcessStatus = styled.div<{ status: string }>`
-  font-size: 8px;
-  padding: 8px;
-  border-radius: 16px;
-`;
-
-const ButtonContainer = styled.div`
-  display: grid;
-`;
-
-const HeaderContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  padding-left: 8px;
-`;
-
-const Title = styled.h4`
-  // Add any styles you want for your title here
-`;
-
-const StyledArrowLeft = styled(ArrowLeft)`
-  color: #FFF;
-`
-
-const TimerDisplayContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  font-size: 8px;
-`;
-
-const TimerDisplay = ({ timers }: { timers: Record<string, number> }) => {
-  return (
-    <TimerDisplayContainer>
-      {timers["startedDownloading"] && timers["finishedDownloading"] ? (
-        <div>
-          Zkey Download time:&nbsp;
-          <span data-testid="download-time">{timers["finishedDownloading"] - timers["startedDownloading"]}</span>ms
-        </div>
-      ) : (
-        <div></div>
-      )}
-      {timers["startedProving"] && timers["finishedProving"] ? (
-        <div>
-          Proof generation time:&nbsp;
-          <span data-testid="proof-time">{timers["finishedProving"] - timers["startedProving"]}</span>ms
-        </div>
-      ) : (
-        <div></div>
-      )}
-    </TimerDisplayContainer>
-  );
-};
