+++ client/src/pages/Swap.tsx
@@ -2,8 +2,7 @@ import React, { useState } from 'react';
 import styled from "styled-components";
 
 import SwapModal from "../components/Swap"
-import { OnRamperProof } from '../components/Swap/OnRamperProof'
-import { OnRamperSubmit } from '../components/Swap/OnRamperSubmit'
+import { OnRamp } from '../components/Swap/OnRamp'
 import { Intent } from "../helpers/types";
 import AccountContext from '../contexts/Account/AccountContext';
 
@@ -18,9 +17,6 @@ export const Swap: React.FC<{}> = (props) => {
    * State
    */
   const [selectedIntent, setSelectedIntent] = useState<Intent | null>(null);
-
-  const [submitOrderPublicSignals, setSubmitOrderPublicSignals] = useState<string>('');
-  const [submitOrderProof, setSubmitOrderProof] = useState<string>('');
 
   /*
    * Handlers
@@ -43,20 +39,13 @@ export const Swap: React.FC<{}> = (props) => {
           onIntentTableRowClick={handleIntentClick}
         />
         ) : (
-        <OnRampProofContainer>
+        <OnRampContainer>
           <Column>
-            <OnRamperProof
-              loggedInWalletAddress={ethereumAddress}
-              setSubmitOrderProof={setSubmitOrderProof}
-              setSubmitOrderPublicSignals={setSubmitOrderPublicSignals}
+            <OnRamp
               handleBackClick={handleBackClick}
             />
-            <OnRamperSubmit
-              proof={submitOrderProof}
-              publicSignals={submitOrderPublicSignals}
-            />
           </Column>
-        </OnRampProofContainer>
+        </OnRampContainer>
       )}
     </PageWrapper>
   );
@@ -70,7 +59,7 @@ const PageWrapper = styled.div`
   justify-content: center;
 `;
 
-const OnRampProofContainer = styled.div`
+const OnRampContainer = styled.div`
   max-width: 660px;
   padding-top: 1.5rem;
 `;
